Fix: Move test helper traits into OpenCFP\Test\Helper namespace
