+++ tests/Domain/Talk/TalkProfileTest.php
@@ -7,7 +7,7 @@ use OpenCFP\Domain\Model\Talk;
 use OpenCFP\Domain\Speaker\SpeakerProfile;
 use OpenCFP\Domain\Talk\TalkProfile;
 use OpenCFP\Test\BaseTestCase;
-use OpenCFP\Test\RefreshDatabase;
+use OpenCFP\Test\Helper\RefreshDatabase;
 
 class TalkProfileTest extends BaseTestCase
 {
