+++ tests/Helper/DataBaseInteraction.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace OpenCFP\Test;
+namespace OpenCFP\Test\Helper;
 
 use Illuminate\Database\Capsule\Manager as Capsule;
 
@@ -8,7 +8,7 @@ trait DataBaseInteraction
 {
     protected function resetDatabase()
     {
-        $this->getCapsule()->getConnection()->unprepared(file_get_contents(__DIR__. '/dump.sql'));
+        $this->getCapsule()->getConnection()->unprepared(file_get_contents(__DIR__. '/../dump.sql'));
     }
 
     protected function getCapsule()
