+++ tests/Infrastructure/Auth/SentryAuthenticationTest.php
@@ -5,7 +5,8 @@ namespace OpenCFP\Test\Infrastructure\Auth;
 use OpenCFP\Infrastructure\Auth\SentryAccountManagement;
 use OpenCFP\Infrastructure\Auth\SentryAuthentication;
 use OpenCFP\Test\BaseTestCase;
-use OpenCFP\Test\DataBaseInteraction;
+use OpenCFP\Test\Helper\DataBaseInteraction;
+use OpenCFP\Test\Helper\SentryTestHelpers;
 
 /**
  * Class SentryAuthenticationTest
